minor cleanup of `get_lambda_max()`
I am using ESOQ-2, on the `adcs` repo, to compute the average of quaternions. In particular, it is used to perform eigenanalysis on the matrix K. The code works, but there is a discrepancy in the computation of the optimal quaternion that I have yet to investigate.

Now that I think about it, I believe I should implement an additional function to compute max eigenvector corresponding to `lambda_max`. This way, the same code can be used in both codebases.